refactor(platform): extract debugger config into config/debugger.rs (#2830)
Move debugger/load-state/tracing/debugger-alpha/breakpoint parsing, the
trace_* config keys, and parse_breakpoint_list/parse_hex_addr into
config/debugger.rs. mod.rs delegates to it. No behavior change.
